refactor: 경로 설정 및 에러 핸들링 수정
